[TMA] fix warning msg, foldable dictionary & supported arch
